+++ src/main/java/com/example/mobileapi/dto/response/OrderResponseDTO.java
@@ -14,7 +14,7 @@ public class OrderResponseDTO {
     private Integer id;
     private Integer customerId;
     private LocalDateTime orderDate;
-    private double totalAmount;
+    private Integer totalAmount;
     private String address;
     private String numberPhone;
     private String status;
